requirements.txt
grpcio
grpcio-tools
requests
beautifulsoup4
kafka-python
pymongo
prisma